moved component header on main-content
