+++ src/js/components/page-content/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import cx from 'classnames';
 import MainNav from '../main-nav';
-import Header from '../main-content__header';
+import Header from '../main-content/header';
 import MainContent from '../main-content';
 import { browserHistory } from 'react-router';
